set components to int data

+++ Assets/Scripts/Game/Components/Wisdom.cs
@@ -2,5 +2,5 @@
 
 public class Wisdom : MonoBehaviour
 {
-    public float Value;
+    public int Value;
 }
